COM-2974 rename CourseList -> LearnerCourses
