+++ src/navigation/Home/index.tsx
@@ -2,11 +2,11 @@ import { Text, View } from 'react-native';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import CatalogIcon from '../../../assets/icons/CatalogIcon';
 import CatalogSelectedIcon from '../../../assets/icons/CatalogSelectedIcon';
-import CoursesIcon from '../../../assets/icons/CoursesIcon';
-import CoursesSelectedIcon from '../../../assets/icons/CoursesSelectedIcon';
+import LearnerCoursesIcon from '../../../assets/icons/LearnerCoursesIcon';
+import LearnerCoursesSelectedIcon from '../../../assets/icons/LearnerCoursesSelectedIcon';
 import ProfileIcon from '../../../assets/icons/ProfileIcon';
 import ProfileSelectedIcon from '../../../assets/icons/ProfileSelectedIcon';
-import CourseList from '../../screens/courses/CourseList';
+import LearnerCourses from '../../screens/courses/list/LearnerCourses';
 import Catalog from '../../screens/explore/Catalog';
 import ProfileDetails from '../../screens/profile/Profile';
 import styles from './styles';
@@ -25,12 +25,12 @@ const catalogIcon = ({ focused }: tabBarProps) => (focused
   </View>
   : <CatalogIcon style={styles.iconContainer} />);
 
-const courseIcon = ({ focused }: tabBarProps) => (focused
+const learnerCoursesIcon = ({ focused }: tabBarProps) => (focused
   ? <View style={styles.iconContainer}>
-    <CoursesSelectedIcon />
+    <LearnerCoursesSelectedIcon />
     <Text style={styles.iconText}>Mes formations</Text>
   </View>
-  : <CoursesIcon style={styles.iconContainer} />);
+  : <LearnerCoursesIcon style={styles.iconContainer} />);
 
 const profileIcon = ({ focused }: tabBarProps) => (focused
   ? <View style={styles.iconContainer}>
@@ -41,9 +41,9 @@ const profileIcon = ({ focused }: tabBarProps) => (focused
 
 const Home = () => (
   <Tab.Navigator screenOptions={{ headerShown: false, tabBarShowLabel: false, tabBarStyle: styles.tabBar }}
-    initialRouteName="Courses">
+    initialRouteName="LearnerCourses">
     <Tab.Screen name="Catalog" component={Catalog} options={{ tabBarIcon: catalogIcon }} />
-    <Tab.Screen name="Courses" component={CourseList} options={{ tabBarIcon: courseIcon }} />
+    <Tab.Screen name="LearnerCourses" component={LearnerCourses} options={{ tabBarIcon: learnerCoursesIcon }} />
     <Tab.Screen name="Profile" component={ProfileDetails} options={{ tabBarIcon: profileIcon }} />
   </Tab.Navigator>
 );
