+++ src/screens/courses/SubProgramProfile/index.tsx
@@ -76,7 +76,7 @@ const SubProgramProfile = ({ route, navigation, setStatusBarVisible, resetCourse
 
   const goBack = useCallback(() => {
     resetCourseReducer();
-    navigation.navigate('Courses');
+    navigation.navigate('LearnerCourses');
   }, [navigation, resetCourseReducer]);
 
   const hardwareBackPress = useCallback(() => {
