+++ src/screens/courses/list/styles.ts
