+++ src/types/NavigationType.ts
@@ -23,7 +23,7 @@ export type RootStackParamList = {
 
 export type RootBottomTabParamList = {
   Catalog: undefined;
-  Courses: undefined;
+  LearnerCourses: undefined;
   Profile: undefined;
 }
